Swaasth_samaj/backend_django/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
PyJWT>=2.8
bcrypt>=4.1
